feat: update Manager entity and create authentication DTOs
- Changed Employee id from int to Long for database consistency

- Added authentication fields to Manager: username, email, password, timestamps

- Created ManagerRegistrationDto with validation annotations

- Created ManagerLoginDto for login credentials

- Added @PrePersist and @PreUpdate lifecycle callbacks

- Fixed all dependent service classes to use Long type

- Maintained Manager inheritance from Employee (JOINED strategy)

Related to Task 2.0 in Manager Authentication System PRD

+++ Project Principle/demo/src/main/java/com/restaurant/demo/model/Employee.java
@@ -11,23 +11,23 @@ import jakarta.persistence.Table;
 @Inheritance(strategy = InheritanceType.JOINED)
 public class Employee {
     @Id
-    protected int id;
+    protected Long id;
     protected String name;
     protected String position;
 
     public Employee() {}
 
-    public Employee(int id, String name, String position) {
+    public Employee(Long id, String name, String position) {
         this.id = id;
         this.name = name;
         this.position = position;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
